requirements.txt
Authlib==1.0.1
cachelib==0.8.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-Caching==1.11.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
python-dotenv==0.20.0
requests==2.28.0
urllib3==1.26.9
Werkzeug==2.1.2
